+++ spec/bs_jwt/authentication_spec.rb
@@ -2,6 +2,24 @@
 
 module BsJwt
   RSpec.describe Authentication do
+    describe '.new' do
+      it 'works with an attributes hash with string keys' do
+        attributes = { 'user_id': '123', 'email': 'test@test.de' }
+
+        authentication = Authentication.new(attributes)
+
+        expect(authentication).to have_attributes(attributes)
+      end
+
+      it 'works with an attributes hash with symbol keys' do
+        attributes = { user_id: '123', email: 'test@test.de' }
+
+        authentication = Authentication.new(attributes)
+
+        expect(authentication).to have_attributes(attributes)
+      end
+    end
+
     describe '#has_role?' do
       subject { build(:bs_jwt_authentication, roles: %w[content_mgr admin]) }
 
@@ -20,6 +38,24 @@ module BsJwt
         let(:authentication) { build(:bs_jwt_authentication, expires_at: 2137.seconds.ago) }
 
         it { expect(authentication).to be_expired }
+      end
+    end
+
+    describe '#to_h' do
+      it 'returns a hash with all attributes of the instance' do
+        authentication = build(:bs_jwt_authentication)
+
+        hash = authentication.to_h
+
+        expect(hash).to eq(
+          'roles' => authentication.roles,
+          'display_name' => authentication.display_name,
+          'token' => authentication.token,
+          'expires_at' => authentication.expires_at,
+          'buddy_id' => authentication.buddy_id,
+          'email' => authentication.email,
+          'user_id' => authentication.user_id
+        )
       end
     end
   end